refactored package and added google-services file
