+++ app/src/androidTest/java/codepath/com/spknwrd/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package codepath.com.instagram;
+package codepath.com.spknwrd;
 
 import android.content.Context;
 import android.support.test.InstrumentationRegistry;
