+++ app/src/main/java/codepath/com/spknwrd/InstaAdapter.java
@@ -1,4 +1,4 @@
-package codepath.com.instagram;
+package codepath.com.spknwrd;
 
 import android.content.Context;
 import android.graphics.drawable.Drawable;
