+++ app/src/main/java/codepath/com/spknwrd/ParseApplication.java
@@ -1,4 +1,4 @@
-package codepath.com.instagram;
+package codepath.com.spknwrd;
 
 import android.app.Application;
