+++ app/src/main/java/codepath/com/spknwrd/ProfileFragment.java
@@ -1,4 +1,4 @@
-package codepath.com.instagram;
+package codepath.com.spknwrd;
 
 import android.content.Context;
 import android.os.Bundle;
