+++ app/src/test/java/codepath/com/spknwrd/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package codepath.com.instagram;
+package codepath.com.spknwrd;
 
 import org.junit.Test;
